feat: Improve output and refactor response parser (#10)
- Improved the output messages for better clarity.
- Refactored the ResponseParser function to ResponseJsonParser for clarity and consistency.
- Moved GetRepoRawName function to gitHelper.go for better organization and code separation.
- Created a new file gitHelper.go to hold git-related helper functions.
- Updated function calls to reflect the changes made.
