+++ internal/pkg/common/common.go
@@ -7,7 +7,6 @@ import (
 	"strings"
 	"bytes"
 	"runtime"
-	"errors"
 )
 
 // RunCommand executes a command-line command with the provided parameters
@@ -109,15 +108,4 @@ func OpenURL(url string) error {
     }
     args = append(args, url)
     return exec.Command(cmd, args...).Start()
-}
-
-// GetRepoRawName returns the raw name of a repository from its full name
-func GetRepoRawName(repository string) (string, error) {
-    repoGitName := strings.Split(repository, ":")
-    if len(repoGitName) < 2 {
-        return "", errors.New("Invalid repository name. " + repository + " could not be parsed properly")
-    }
-	// Removing '.git' from name
-	repoName := strings.ReplaceAll(repoGitName[1], ".git", "")
-    return repoName, nil
 }
